Move `OpenDeviceArgs` and `ActiveDeviceState` into `device.py`
