+++ tests/gui/hardware_set/conftest.py
@@ -5,7 +5,8 @@ from pathlib import Path
 
 import pytest
 
-from finesse.gui.hardware_set.hardware_set import HardwareSet, OpenDeviceArgs
+from finesse.gui.hardware_set.device import OpenDeviceArgs
+from finesse.gui.hardware_set.hardware_set import HardwareSet
 
 _HW_SETS = (
     HardwareSet(
