+++ tests/gui/hardware_set/test_hardware_sets_combo_box.py
@@ -6,7 +6,8 @@ from unittest.mock import MagicMock, Mock, PropertyMock, call, patch
 
 import pytest
 
-from finesse.gui.hardware_set.hardware_set import HardwareSet, OpenDeviceArgs
+from finesse.gui.hardware_set.device import OpenDeviceArgs
+from finesse.gui.hardware_set.hardware_set import HardwareSet
 from finesse.gui.hardware_set.hardware_sets_combo_box import HardwareSetsComboBox
