+++ tests/gui/hardware_set/test_hardware_sets_view.py
@@ -9,10 +9,9 @@ from unittest.mock import MagicMock, Mock, PropertyMock, call, patch
 import pytest
 
 from finesse.device_info import DeviceInstanceRef
+from finesse.gui.hardware_set.device import ActiveDeviceState, OpenDeviceArgs
 from finesse.gui.hardware_set.hardware_set import (
-    ActiveDeviceState,
     HardwareSet,
-    OpenDeviceArgs,
 )
 from finesse.gui.hardware_set.hardware_sets_view import (
     ActiveDeviceProperties,
@@ -244,7 +243,7 @@ def test_update_control_state(
     (((), range(2), range(2)), (range(2), range(2), ()), ((0,), range(2), (1,))),
 )
 @patch("finesse.gui.hardware_set.hardware_sets_view.settings")
-@patch("finesse.gui.hardware_set.hardware_set.open_device")
+@patch("finesse.gui.hardware_set.device.open_device")
 def test_connect_btn(
     open_mock: Mock,
     settings_mock: Mock,
